changes as per review comments and varuns comments on stats blob

+++ callstats-java-sdk/src/main/java/io/callstats/sdk/data/StreamStatsData.java
@@ -32,7 +32,7 @@ public class StreamStatsData {
 		this.jitter = jitter;
 	}
 	
-	public StreamStatsData(int rtt, long packetsSent, long bytesSent, float jitter) {
+	public StreamStatsData(int rtt, long packetsSent, long bytesSent, double jitter) {
 		super();
 		this.rtt = rtt;
 		this.packetsSent = packetsSent;
